server/requirements.txt
anyio==3.6.2
asttokens==2.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
black==22.10.0
boto3==1.26.3
botocore==1.29.3
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
debugpy==1.6.3
decorator==5.1.1
dnspython==2.2.1
entrypoints==0.4
executing==1.2.0
Flask==2.2.2
idna==3.4
ipykernel==6.17.0
ipython==8.6.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
jupyter_client==7.4.4
jupyter_core==4.11.2
Markdown==3.4.1
markdownify==0.11.6
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mypy-extensions==0.4.3
nest-asyncio==1.5.6
packaging==21.3
parso==0.8.3
pathspec==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.2
prompt-toolkit==3.0.32
psutil==5.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.2
Pygments==2.13.0
pymongo==4.3.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pyzmq==24.0.1
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.0
starlette==0.20.4
tomli==2.0.1
tornado==6.2
traitlets==5.5.0
typing_extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
